Add matrix * float operators to the math library
As per earlier code review feedback, this adds matrix*float versions of Multiply and operator* to the Matrix3x2 and Matrix4x4 types.

Unlike the vector types, I did not include float*matrix version of operator*, or matrix divided by float operations.  This makes the matrix type consistent with Quaternion, and I think is appropriate as the scenarios for multiplying matrix by scaler are not common enough to justify expanding out this full set of minor variants.