Fix midl include path.
The $(ProjectDir) was unnecessary (these directories are implicitly relative to the directory containing the vcxitems file).  In addition, $(ProjectDir) was getting set to different cases depending if the build is initiated from within Visual Studio or from within the command-line.  This was erroneously triggering the "the command-line changed so I'll rebuild everything" path in msbuild.

Options to fix here were to force the case on $(ProjectDir) -- or just remove it since it is redundant.