CanvasBrush no longer has a "Device" property.
This is expected to extend to most other resources created against a device.

The device property proved to be problematic since ID2D1SolidColorBrush has no way to obtain the ID2D1Device that the brush was created against.  This meant that a CanvasSolidColorBrush created from a ID2D1SolidColorBrush had no way to return the Device property safely.

One option here would be to have it return null in this case, but this would mean that anyone trying to use the property would need to check for null and arrange for the device to be available some other way.  This is effectively the same as removing the property.  Leaving it there would be baiting a trap.

Also there were general efficiency concerns about having any extra back pointers than are strictly necessary.

And finally, there's a suspicion that if you're about to get the Device from a CanvasSolidColorBrush then the chances are that you're about to do something that would be considered an anti-pattern.