CanvasControl - improve resizing behavior
Set the imagecontrol's Stretch mode to None.   This prevents the control from resizing when we set a different Source on it.

Create the Image control once on startup (#1770).  This avoids us needing to change the visual tree whenever there's a resize.

Filter SizeChanged events to cater for receiving SizeChanged events.  This ensures that in a simple startup case we only fire the Drawing event once.