NuGet packages reworked to work correctly with designer and windows phone apps
We now tell nuget about the winmd file we're adding so that the designer can pick this up.

To do this we drop copies of the x86 build into the 'win' and 'wpa' directories.  These are only used by the designer.

We also now drop build scripts in 'win', 'wpa' and 'native' directories.  This simplifies the scripts a little and also indicates to VS that we support win, wpa and native 'frameworks'.