Documentation links out to MSDN for WinRT types referenced by our API
Sandcastle has only limited support for external links to types that aren't in the
assembly being documented. So I can't just point it straight at MSDN for WinRT types like
SurfaceImageSource that our API makes use of.

The right longterm fix is to write a Sandcastle plugin to do better reference resolution,
but I don't have time for that now.

An easy option is just to exclude these types, in which case they show up not hyperlinked
at all, but that is a pain especially for things like CanvasTextFormat that uses lots of
XAML enums.

So instead, we generate empty versions of these types in our Sandcastle doc, and give
those a summary comment that links to the real MSDN doc page. Requires two clicks to get
there instead of one, but I figure it's better than nothing.

To do this, we now generate XML doc comments for these placeholder types in order to
provide the MSDN link, and also namespace summary documentation, which is output as a
Sandcastle project fragment as opposed to doc XML (not sure why doc XML can't handle
namespaces, but it can't).