DocDiff tool validates that our documentation properly matches the latest API surface
This runs as part of the doc build and compares the placeholder XML docs created by
ExtractAPISurface tool against the real ones that are checked into the docsrc folders. It
warns if new APIs are missing their documentation, or old APIs have been removed but the
stale documentation for them is still there.

If documentation is missing, a placeholder XML file is written to obj\docs\MissingDocs.xml.
This is formatted in such a way that its contents can be pasted directly into the
appropriate docsrc files, as a quick way to get started filling in the tags with real doc
content.

Used this to remove a couple of stale APIs from the current docsrc XML files.

Also updated the ExtractAPISurface tool to include doc comments for every member value of
enums.