Rename implementation namespace: canvas -> ABI::Microsoft::Graphics::Canvas
Why?

When MIDL sees a runtime class, it thinks this goes in ABI::Foo::Bar::Blah
namespace - in our case this is ABI::Microsoft::Graphics::Canvas. It generates
code putting lots of useful interface definitions in that namespace, plus a
forward declaration for the runtime class. We ignored that forward declaration
and proceeded to write all our code in the "canvas" namespace instead. This
mostly works, except:

a) It sometimes causes confusion when both namespaces are in scope, as the
   compiler can pick up the MIDL generated forward declaration instead of the
   actual implementation class. We had lots of using statements and explicit
   namespace qualifiers in test code to overcome that, which are no longer
   needed.

b) More importantly, when MIDL sees a generic that is specialized on a runtime
   class (collection of classes, or async operation that returns a runtime
   class) it generates a bunch of template specialization magic to make
   everything wire up together. If you then go an use a runtime class from a
   different namespace than the one MIDL expected, these specializations don't
   work, hair is torn out, and compile error message are mysterious and hard to
   figure out.

This stuff is not configurable in MIDL, so putting our implementations in the
expected namespace is the only way to make these generic types work correctly.