Rename CanvasBitmap.SizeInDips to CanvasBitmap.Size
This:

a) makes it consistent with XAML where, so "Control.Size" and "CanvasBitmap.Size" use the same units.

b) guides developers to use DIPS.  Previously the choice between "SizeInDips" and "SizeInPixels" made DIPS and Pixels seem as if they're equally preferred, leading to a trap of using pixels.  Since pixels happen to equal dips in many cases this opens up for surprises when run on displays where they don't.