CanvasStrokeStyle.DashCap defaults to Square rather than Flat
Dots are not visible when DashCap is Flat (they have zero size),
which can be a confusing result. Defaulting this to Square makes them
visible by default.

Also updated documentation to explain the behavior of DashCap = Flat,
and added text explaining it to the strokes demo in ExampleGallery.

This addresses https://github.com/Microsoft/Win2D/issues/14