CanvasControl now tested via events rather than directly calling the event handlers
Previously, all the tests passed even if CanvasControl didn't register the Loaded or SizeChanged events.

Now the tests fail if the control doesn't respond to these events.

In addition some of the resizing tests were explicitly requesting redraws; now these tests use the control's built-in redraw-on-resize behavior (and will fail if this behavior breaks).

Also replaced a magic '1000' with a variable.