Add Game of Life to Example Gallery
I intend to use this as a testbed to convince myself we really have a buttoned
down story re. high DPI. It stresses this area in several different ways:

- Draws to rendertargets which don't match the display size, then does an aspect ratio preserving scale to fit the screen
- Rendertargets are accessed both as drawing surfaces and via GetColors/SetColors
- Uses pointer input to modify pixels in the rendertarget, and expects that to line up correctly
- The algorithm is very sensitive to filtering - won't work if any unexpected scales get inserted behind the scenes

So far I haven't done anything special to handle DPI - I've written this in a
naive way assuming that pixels and DIPs are the same thing, i.e. pretending I
don't know there could be dragons here. Looking at the behavior of this +
burning text + the existing effect demos, we definately don't have an easy "it
just works" DPI story in place yet. Next step is to understand exactly what's
going wrong and come up with a plan to fix it.

I think it's pretty cool that image effects can implement a Turing complete
cellular automaton that runs on the GPU in so few lines of code!