Change how DPI works when loading bitmaps
DPI metadata in the source image file is now ignored (because it turns out this
is wrong too often to be relied on in image files found out in the wild, and to
match the behavior of most other APIs and rendering engines).

Instead, we add CanvasBitmap.LoadAsync overloads that explicitly specify what
DPI to load at. If the DPI is not explicitly specified, it defaults to 96.

Also removed a couple of unnecessary method overloads (CanvasRenderTarget
constructor that took a Size struct, and bitmap + rendertarget
CreateFromDirect3D11Surface methods that took DPI but not alpha behavior).