Cosmetic improvements to make Example Gallery burning text and effect demos look good on Phone
Added dynamic scaling on a few things that were otherwise too big to fit the screen.

Dropped the text overlay from the effect demo when running on screens too small
to have room for it.

Fixed some stuff that was slightly off center, which wasn't noticeable on a
large display but got ugly as the screen size decreases.

Optimized how the luminance to alpha demo used turbulence, which was too slow on phone HW.