C++ numerics idiomaticifation
Rewrite the C++ numerics implementation in idiomatic C++, replacing the
previous transliteration of the .NET version. This is not yet SIMD optimized,
just more concise and easier to read/maintain. Now relying on the compiler to
inline things so complex methods can be written in terms of simpler ones.

Also found a longstanding cornercase bug in Matrix4x4.Decompose, which Win2D
inherited from XNA. Fixed this in both the C++ and C# versions, and updated
tests to match.