SizeInPixels properties return integers rather than floats
Most Win2D sizes and positions are in device independent units, which are of
type float, but pixel units are integers, so the SizeInPixels properties should
return an integer type.

Added a new BitmapSize struct to make this possible.

Also noticed that CanvasImageSource was missing any way to query its size, so
added Size and SizeInPixels properties to it.

Doc improvement