CanvasAnimatedControl handles the control being zero size
This now keeps the update/render thread going, but if the control is zero sized
it only calls Update while skipping Draw. It also properly handles the
transition from having a swapchain to being zero size, which needs to
temporarily flip back to the UI thread in order to set the SwapChainPanel to
use a null swapchain.