Memory leak fix: ResourceManagers now clean up their ManagerHolder
We use CoreApplication.Properties to stash weak references that can later be used to look
up our manager instances. This was mostly fine, but the weak reference had to be wrapped
in an IInspectable in order to add it to the property bag, producing a tiny ManagerHolder
helper. We leaked these.

The leak was causing AVs during unit test teardown (after the test already recorded a
pass status, so only visible in the debugger) because CoreApplication would call Release
on our ManagerHolder after the Win2D DLL was unloaded.

With the new lifespan tracker stuff in place, the leak caused a large number of unit test
failures.

Fix is for ResourceManager to remove its ManagerHolder from CoreApplication.Properties
whenever the manager gets destroyed.