Memory leak fix: add virtual destructors to types that were missing them
Most of these omissions were probably benign, but at least one (in
RecreatableDeviceManager.h) was causing a significant leak.

Also turned on the compiler warning to make sure we do this everywhere from now on.
Annoyingly, that warning only fires for classes where the compiler generates a
non-virtual destructor, and not for abstract interfaces where it's most important!