Fix hang when CanvasTextFormat uses a custom font from the UI thread
When the CanvasTextFormat realizes the IDWriteTextFormat it uses GetFileFromApplicationUriAsync() to determine the actual pathname for the custom font file.  In the completion callback for this operation, an event is set, and the calling code waits for the event.

When the realization happens on a non-UI thread (such as the CanvasAnimatedControl's game loop thread) then the callback executes on another threadpool thread.

However, when it happens on the UI thread, the callback ends up being scheduled to run on the UI thread - but that has been blocked waiting for the callback to complete --> deadlock!

The fix here is to make the callback agile (by using FTMBase).  This will cause the callback to be scheduled on another thread.