Better opacity mask documentation (github #61)
Add cross links between docs about layers + the FillRectangle and FillGeometry overloads
that take an opacity brush (as these are alternative ways of applying opacity masks).

Add wording to clarify it is the alpha of the opacity mask that matters (not its RGB).

Use the term "opacity mask" in more places, as experience has shown this is what
customers are looking for, yet that wording was previously absent from our docs :-(