Win2D can now be consumed by a managed windows runtime component
Explicitly remove reference to platform.winmd when building managed projects

This reference is unfortunately pulled in when we reference Microsoft.VCLibs.  It conflicts with windows.winmd in a way that the managed toolchain does not understand.

It seems that the only right thing to do is to remove the reference.