CanvasImageBrush bugfixes
Fixes two related issues:

1) Accessing base ICanvasBrush properties (Opacity or Transform) on an image brush was
triggering unwanted brush state validation, so would throw if you did this before setting
the source rect.

2) Missing conversion logic in the Image setter - if you assigned a bitmap to a brush
that was currently in image mode but had no source rect set, it didn't realize it could
convert back to a bitmap brush (conversion would happen if you assigned the bitmap and
then set source rect to null, but not when setting these properties in the opposite
order).