Fix deadlock around multithreaded use of SaveAsync and other D2D methods (github #110)
The issue here is that when the SaveAsync completion handler is called from a thread other than the original thread a deadlock can occur where the second thread is stuck waiting for the async action to be destroyed before it can take out the D2D lock.

The fix is to ensure that SaveAsync takes out and releases the lock from within a single thread of control.