RunOnGameLoopThreadAsync work is canceled when the control is not loaded
When the control is unloaded any pending RunOnGameLoopThreadAsync work is canceled.

If RunOnGameLoopThreadAsync is called while the control is not loaded then the work is immediately canceled.

This avoids the situation where an app is waiting for an IAsyncAction to complete that'll never have a chance to complete because the game loop thread isn't running.