Remove over-eager ResourceTracker assert
This could fire erroneously if a Win2D object constructor threw after its base
ResourceWrapper was constructed but before the instance was registered with
ResourceTracker. In that case the ResourceWrapper destructor still tries to remove itself
from the ResourceTracker, which would assert because it did not recognize the object
being removed.

This situation is actually fine, so now we just silently do nothing.