Move bitmap resource creation to CanvasDevice
This removes the places where CreateDeviceContext() was being used
just to create a device context to create resources.  The pattern here
is to use CanvasDevice to create the resources (allowing CanvasDevice
to own the policy of how these resources are created).

CreateDeviceContext() is renamed to CreateDeviceContextForDrawingSession()
to make the purpose of this method clearer.

GetResourceCreationDeviceContext() remains to allow CanvasTextFormat to
pass a device context to GetD2DBrush.  This is used to create the
DPI compensation effect.