Remove TODO from CanvasDrawingSession::GetColorBrush
This TODO was suggesting that we pool / reuse the
SolidColorBrush that CanvasDrawingSession uses to support
the various overloads that take a Color.

This was based on the assumption that we'd be taking steps to try and
reuse device contexts between drawing sessions - which we've decided
isn't necessary.  There haven't been any indications that creating
this solid color brush once per drawing session is causing any
performance issues.

Also, since a CanvasDrawingSession is not intended to be shared between
threads, there are no thread safety considerations here.