Make CanvasSwapChain threadsafe
Anything that uses m_dpi in conjunction with obtaining the size from
the swapchain needs to operate under a lock, since these two values are
updated together.

CanvasSwapChain is already using the D2DResourceLock to serialize access
to the swapchain, so we can hijack this to serialize access to the
swapchain and m_dpi.

At the same time we can smooth out some of the rough edges - eg calling
ResizeBuffersWithSize would have previously created two D2DResourceLocks.