Initial effect interop implementation
With this change, effects are tracked by the interop system and can be interopped in both
directions between Win2D and D2D. This fixes issues with reading the value of an
ImageBrush.Image property that has been set to an effect.

However, the effect state is not yet fully synced between Win2D and D2D:

- When calling GetWrappedResource, Win2D overwrites all D2D properties with its own state
- D2D property values are snapshotted when GetOrCreate first creates a Win2D wrapper
- D2D property changes to an already-wrapped resource are not reflected on the Win2D side

Full 2-way state synchronization is coming later.