New mechanism for ResourceWrapper to convert 'this' to the outer IInspectable
To add itself to the interop table, ResourceWrapper needs to look up the outer
IInspectable for whatever runtime class has mixed in the ResourceWrapper
instance.  It is important to use the right IInspectable (of which there are
several copies due to multiple inheritance) because when the ResourceWrapper
constructor executes, things are not yet initialized far enough for QI or
AsWeak to work from any of the other versions.

The previous implementation worked for some types but not all (it can't handle
CanvasEffect).   New approach is more robust, and also has the advantage that
it lets me move the AsWeak operation from ResourceWrapper into ResourceManager
(where it belongs).