Modify IBasicEffect implementations to return IsReadOnly = false
IsReadOnly = true means "this effect does not process the image at all"
and will cause the effect system to just copy the input to the output.
Then, when we do try and modify the output we end up fighting with it
leading to odd timing issues.

IsReadOnly = true is intended for analysis-type effects.  All the effects
in example gallery really intend IsReadOnly to be false.