Make print system callbacks block until the work has completed
Paginate, MakePage and MakeDocument are called by the printing system
on a threadpool thread.  We do the actual work asynchronously on the UI
thread (with potential deferrals).  However, the only mechanism to
report failures back to the printing system is by the HRESULT returned
by these callback methods.  Therefore, we block these methods from
returning until the underlying task has completed.

We disable this behavior in our main tests since we want to avoid
dealing with multithreading issues during tests.

Specific tests for DeferredTask validate that we capture exceptions and
pass them through in a way that Paginate, MakePage and MakeDocument
can convert to hresults.