Include a hard-coded version of the numerics intellisense docs for Win81
This file was previously generated as part of the documentation for the
Microsoft.Graphics.Canvas.Numerics.dll.

However, now that we've switched to Windows 10 as the focus for
documentation this file is no longer generated as we don't
ship these APIs with Win2D.

Fortunately, we're very unlikely to need to modify this file so this
should be a reasonable workaround.