Doc validation tool bugfix
Validation for cref links with O: overload groups as their target wasn't handling methods with no args.

It was looking for "O:methodName(" when counting that more than one such method exists (because if there is
only one method, it's not valid to link to that via O: syntax) but this check was failing if there was only
one method with arguments plus a second without, in which case the latter method has no trailing "(" in its
name.