Call RoClearError when handling device lost errors
This prevents the error state from staying around, which can cause
problems if there's a later error that doesn't call RoOriginateError.

XAML uses the error state to override the error code.  The result here
is that the UnhandledException handler will get called with the
device lost error state, rather than the real error state that was unhandled.

This is related to issue #194 where it looked like a device lost exception
was being unhandled, when instead it was another error.