Fix test break introduced by the new *_MaxSizeError external tests
When running tests on a WARP device, MaximumBitmapSizeInPixels is much larger than on any
real HW GPU. Large enough that when the tests declare:

	auto badSize = MaximumBitmapSizeInPixels + 1;

and then cast this to a float to pass to Win2D constructor APIs, the resulting float
rounds back downward and ends up the same as the legit max size!

Fix is to more carefully choose a bad value that is accurately representable using 32 bit
float as well as int parameter types.