Change when CanvasAnimatedControl calls WaitForVerticalBlank
Previously, WaitForVerticalBlank was always called at the beginning of
each tick.  If the previous frame was delayed for some reason (eg the
Present() blocked) past the VBlank, then this wait would compound the
problem, resulting in a missed frame.

This version only calls WaitForVerticalBlank if we did not present a
frame.

Generally, Present() blocks anyway to prevent the CPU getting
too far ahead of the GPU, so the CPU is not busy looping in this case.

However, in the cases where Present() does return early enough that
it isn't time to do a tick then it is worth waiting for the vblank to
avoid busy waiting.