Fix CanvasBitmap::SaveAsync to support pixel formats other than B8G8R8A8UIntNormalized
This change makes SaveAsync use IWICImageEncoder, rather than directly
locking the surface and passing the pixel data to WIC.

IWICImageEncoder will, if necessary, convert the bitmap to
B8G8R8A8UintNormalized before saving it.

Note that a bug (6184116) IWICImageEncoder means that the debug layer
will issue a message / debug break whenever the a bitmap is saved. For
this reason the tests disable the debug layer when testing SaveAsync.