Fix race condition in GameLoopThreadTests
Our stub ProcessEvents implementation was waiting on a condition variable that might have
already been signalled, but this is not how condition variables work. If notify is called
before wait, there's nothing to notify so that's a no-op. Therefore we should only wait
if there isn't already work queued up.