Convert test.external.uap to use CppUnitTestFramework.Universal
This was previously using the non-universal version of the test framework, which somehow
managed to work ok with VS 2015 Update 1, but is not the right thing and broke with
Update 2.

Also reorganized the directory structure to keep things clear as this test project now
includes some UAP-only app files. test.external is split into Shared, UAP, Windows, and
WindowsPhone sub folders, just like we already did for the samples and test.managed.