Update print preview when switching printers
If you switch printers the print preview doesn't update until you change
another property like Orientation. This is noticeable when switch from
something like print to PDF which has no margins to a physical printer
that does. (bug #292)

Fix is to listen to OptionChanged and invalidate the preview when an
event with a null OptionId is raised.