Fix github #319: Creating CanvasAnimatedControl in code sometimes hangs
Turns out that when a custom XAML control is used as an inline element within a Paragraph
(via InlineUIContainer), its Loaded event gets raised twice in a row, followed by a pair
of Unloaded when the control is later removed.

Win2D CanvasAnimatedControl did not expect this, and did not handle it well.

The fix is to maintain a count of how many times the control has been loaded, and only
actually do init or cleanup work on the outermost pair.