Rename Win10NovUpdate documentation tag to Win10_10586
"NovUpdate" terminology becomes unclear now we're no longer in the same year when that
update was released. Looks like the OS build # is the mostly widely used way to identify
a specific revision of Win10, and developers will already be familiar with that from the
Visual Studio SDK selection UI, so let's use it in the Win2D docs as well.