CanvasRegionsInvalidatedEventArgs should not be activatable
There is no activation factory for this type, so it was never actually possible for an
app to construct one, but the IDL incorrectly marked the type as activatable. This meant
the docs showed it as having a default constructor. Code attempting to call that would
compile but then fail at runtime.