Make sure CanvasControl::Changed always executes on the UI thread
The Changed implementation calls a couple of XAML APIs that have thread affinity, so
will fail if called on any other thread.

- Multithreaded Win2D app shares a device across more than one thread
- Hit TDR
- CanvasDevice.DeviceLost event is raised on whichever thread notices the TDR first
- BaseControl.OnDeviceLost for other controls executes on the thread of the event sender (which may not be their UI thread)
- OnDeviceLost calls Changed
- CanvasControl.Changed calls IsWindowVisible
- This throws "The application called an interface that was marshalled for a different thread."

The solution is for Changed to marshall itself back onto the UI thread if invoked
anywhere else. CanvasAnimatedControl and CanvasVirtualControl already do that - it is
only CanvasControl where this was a problem.

Also merged the two different ways that BaseControl had to query window visiblity.
IsWindowVisible polled on every call, which is always correct but has UI thread affinity.
IsVisible cached a state and listened for change events, which allows querying from any
thread but returned wrong data if the window was already hidden when the control was
first created. Now we have a single implementation that queries on load, then listens
to subsequent change events.

Fixes github #445 - "Device lost handling can lead to calling IWindow::get_Visible on the wrong thread"