Allow use of CanvasGeometry and CanvasPathBuilder without a CanvasDevice
This change allows a null resource creator to be specified when creating geometry objects.  The resulting geometry can still be manipulated in all the normal ways, but cannot be drawn via a CanvasDrawingSession or converted into a CanvasCachedGeometry.  Attempting to do so will fail with the existing error mechanism that handles resources from mismatched devices.

The motivation for this is to allow use of Win2D for geometry manipulation in situations where the cost of creating D3D+D2D devices is undesirable.