Add ARM64 entries where ARM entries are
This change adds ARM64 entries where ARM(32) entries already exist.

In the case of certain test configurations, ARM is set to use the x86
configuration so ARM64 imitates this.

Since winrt.test.managed.uap.csproj is not compatible with .NET Native
and the ARM64 UWP does not support .NET other than Native, it is
omitted.