Merged PR 16988: Fix AnyCPU compilation
Previously the projection project was converted over to not use AnyCPU anymore. This was because there's a bug where the WindowsAppSdk bootstrap dll tries to deploy an AnyCPU version (which doesn't exist) and fails.

Not using AnyCPU causes issues down the line with adding more tests, so this change restores back to using AnyCPU. But it also addresses the bootstrap dll bug more directly, and works around it to deploy x86.