Merged PR 22516: Disable AppX WinMD caching for consuming apps
Due to a bug in the WinAppSDK VS extension, the Activatable Class entries are not being properly generated. This results in incremental builds that consume Win2D to fail. This workaround will fix the issue in consuming apps until the root bug is fixed, which is slated for the WinAppSDK 1.2 timeframe.