Merged PR 23311: Adding telemetry to the UWP version of Win2D
Essentially the same PR as the [WinUI3 telemetry addition one](https://mscodehub.visualstudio.com/Win2D/_git/Win2D/pullrequest/22459), but without the build pipeline updates built in, and with a new section added to the telemetry data that specifies that the source of the event was the UWP flavor of Win2D.

This PR contains a stubbed version of the telemetry headers. After merging, I will add the proper commands to the ADO build pipelines to replace those with the internal versions.