Merged PR 23861: Add CanvasDevice.IsDeviceLost() and CanvasDevice.GetDeviceLostReason() APIs
## Background

The way Win2D exposes information related to devices being lost is quite limited, and the exposed API (`IsDeviceLost`) is really only meant for a very specific scenario (ie. exception handlers) and not providing much information otherwise. In the Store, we'd like to add some telemetry to track device lost events, but currently there's no way for us to get the reason why a device was lost. The event args have no info whatsoever, and `CanvasDevice` has no APIs to actually query the device lost reason. This PR addresses this 🙂

There's actually a longer explanation, I've included this in the comments in the PR as well.

## API changes

This PR adds two new APIs to `CanvasDevice`:

```csharp
class CanvasDevice
{
    bool IsDeviceLost();
    int GetDeviceLostReason();
}
```

Additional notes (again you can also find more in the comments in the code):
- `IsDeviceLost()` is an overload for the already existing `IsDeviceLost(int)` method. In particular, this doesn't require callers to pass in an HRESULT, but simply tells them whether the device is lost or not.
- `GetDeviceLostReason()` returns the (real) reason why the device was lost. The return value is an `int`, which matches the already existing `int hresult` parameter for `IsDeviceLost(int)`. Additionally, this not being an `HRESULT` means callers will be able to correctly get the real value with no marshalling (whereas if it had been an `HRESULT`, it'd have been exposed as an `Exception` object to C#, which would've been bad).