Merged PR 23817: Add WIN2D_GET_DEVICE_ASSOCIATION_TYPE enum
> **NOTE:** related issue is https://microsoft.visualstudio.com/OS/_workitems/edit/42913458.

## Background

In !23226, we introduced the new `ICanvasImageInterop` type to enable external effects, and we've since started using this in the Store as well to power some of our new UX designs. We recently discovered an issue that causes a crash in the Store following a redraw after a device lost, if the same effect graph containing at least an external effect used as source for a Win2D effect is used. I investigated this and came to the conclusion that this was due to a limitation in the design of `ICanvasImageInterop`, which this PR fixes.

The issue is: `CanvasEffect` has logic that is used when setting an input, which is as follows:
- If the input is an `ICanvasResourceWrapperWithDevice`, get the device.
  - If the device doesn't match the realization device, throw.
- Otherwise, just keep realizing the effect normally.
  - This means if the source is currently realized on a different device, it'll just "reload" and work fine.

This second scenario wasn't supported for `ICanvasImageInterop`, and it's what was causing the crash we saw in the Store. Basically, you had an effect, drawn it once, then changed device, drawn it again, and this time Win2D would throw because it would see that effect was realized on a different device (the previous one). The core of the issue is that there was no way for an `ICanvasImageInterop` implementation to tell Win2D whether the returned device was a realization one (which can be changed), or a creation one, which should be handled differently. This is also the reason why the code paths for internal and external effects weren't equivalent.

## API changes

This PR makes the following changes to the public header:

```cpp
 // The type of canvas device returned by ICanvasImageInterop::GetDevice, describing how the device is associated with the canvas image.
typedef enum WIN2D_GET_DEVICE_ASSOCIATION_TYPE
{
    WIN2D_GET_DEVICE_ASSOCIATION_TYPE_UNSPECIFIED = 0,          // No device info is available, callers will handle the returned device with their own logic
    WIN2D_GET_DEVICE_ASSOCIATION_TYPE_REALIZATION_DEVICE = 1,   // The returned device is the one the image is currently realized on, if any
    WIN2D_GET_DEVICE_ASSOCIATION_TYPE_CREATION_DEVICE = 2,      // The returned device is the one that created the image resource, and cannot change
    WIN2D_GET_DEVICE_ASSOCIATION_TYPE_FORCE_DWORD = 0xffffffff
} WIN2D_GET_DEVICE_ASSOCIATION_TYPE;
```

```diff
class __declspec(uuid("E042D1F7-F9AD-4479-A713-67627EA31863"))
ICanvasImageInterop : public IUnknown
{
public:
-    IFACEMETHOD(GetDevice)(ICanvasDevice** device) = 0;
+    IFACEMETHOD(GetDevice)(ICanvasDevice** device, WIN2D_GET_DEVICE_ASSOCIATION_TYPE* type) = 0;
};
```

This allows external effects to be able to also indicate the association type that they have with the returned canvas device. This way, Win2D can use this information to perform the same exact check i...