Merged PR 24053: Minor fixes to public C exports, add managed test
This PR is a small follow up to the previous ones adding C exports. It includes:
- Replaced `nothrow` with `noexcept` for all public C APIs
- Use new `WIN2DAPI` define for `dllexport`
- New managed test for all C exports
- Add explicit error for published header outside of C++
- Fixed exported symbols for C functions on x86

Specifically, with respect to the first two points, I've chatted with Aaron Robinson (.NET interop team lead) about it to get more insights, and he mentioned that:
> "The `nothrow` is also unnecessary in this case.
> It doesn't mean anything in practice. If you throw an exception across an extern "C" you are saying the API is a C API and that means UB. MSVC won't even enforce it. Instead if you really want to be clear about intent, use noexcept. That will force MSVC with insert a std::terminate in the case when an exception attempts to escape.
> the `nothrow` is a lot of ceremony for nothing.
> [...] I prefer `noexcept`."

So I've made that switch here. Additionally, with respect to `dllexport`, he pointed out how the fact it was hardcoded in the header meant that C++ devs referencing that header would've also defined that exported function themselves, instead of just consuming that. To fix this, I've added a new `WIN2DAPI`, which is defined according to whether the header is being build from Win2D or externally, and that toggles between `dllexport` and `dllimport`. Lastly, I've added managed tests for all C exports, to double check they're correctly exported on all CPU architectures too.

For the explicit error, still from Aaron:

> "If someone includes the header file as is, they are going to get a syntax error because namespace isn't a valid C keyword. Providing that gives them a clear error that lets them know it was designed for their use case - including in a C translation unit. It isn't required, but when I "ship" a .h file that isn't designed to work in C, I generally add it as a good citizen."

And he added that especially for official headers it's just good practice to include it.